+++ pytsite/metatag/__init__.py
@@ -1,10 +1,10 @@
 """PytSite Meta Tags
 """
-from ._api import dump, dump_all, get, reset, t_set, rm
-
 __author__ = 'Oleksandr Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
+
+from ._api import dump, dump_all, get, reset, t_set, rm
 
 
 def _init():