+++ pytsite/pip/__init__.py
@@ -1,6 +1,5 @@
 """PytSite pip Support
 """
-
 __author__ = 'Oleksandr Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'