+++ pytsite/plugman/_error.py
@@ -1,6 +1,5 @@
 """PytSite Plugin Manager Errors
 """
-
 __author__ = 'Oleksandr Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'