+++ pytsite/reload/__init__.py
@@ -1,11 +1,11 @@
 """PytSite Reload Package
 """
-# Public API
-from ._api import RELOAD_MSG_ID, reload, on_before_reload, on_reload, set_flag, get_flag
-
 __author__ = 'Oleksandr Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
+
+# Public API
+from ._api import RELOAD_MSG_ID, reload, on_before_reload, on_reload, set_flag, get_flag
 
 
 def _init():